+++ src/router/AppRouter.tsx
@@ -1,5 +1,6 @@
-import { Route, Routes } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
+import TabBar from '@/components/TabBar/TabBar';
 import Developer from '@/pages/Developer';
 import Main from '@/pages/Main';
 import MyPage from '@/pages/MyPage';
@@ -7,14 +8,21 @@ import Pot from '@/pages/Pot';
 import TeamShareListSearch from '@/pages/TeamShareListSearch';
 
 const AppRouter = () => {
+	const isTabBarVisible = (path: string) => {
+		return ['/', '/search', '/mypage', '/pot'].includes(path);
+	};
+
 	return (
-		<Routes>
-			<Route path="/" element={<Main />} />
-			<Route path="/search" element={<TeamShareListSearch />} />
-			<Route path="/mypage" element={<MyPage />} />
-			<Route path="/pot" element={<Pot />} />
-			<Route path="/developer" element={<Developer />} />
-		</Routes>
+		<BrowserRouter>
+			<Routes>
+				<Route path="/" element={<Main />} />
+				<Route path="/search" element={<TeamShareListSearch />} />
+				<Route path="/mypage" element={<MyPage />} />
+				<Route path="/pot" element={<Pot />} />
+				<Route path="/developer" element={<Developer />} />
+			</Routes>
+			{isTabBarVisible(location.pathname) && <TabBar />}
+		</BrowserRouter>
 	);
 };
